Refactor host info models
