+++ epp-client/src/epp.rs
@@ -5,12 +5,10 @@ pub mod request;
 pub mod response;
 pub mod xml;
 
-pub use request::host::info::*;
 pub use request::host::update::*;
 pub use request::message::ack::*;
 pub use request::message::poll::*;
 
-pub use response::host::info::*;
 pub use response::host::update::*;
 pub use response::message::ack::*;
 pub use response::message::poll::*;
