+++ epp-client/src/epp/request/host.rs
@@ -1,4 +1,3 @@
 //! Types for EPP host requests
 
-pub mod info;
 pub mod update;
