+++ epp-client/src/epp/response/host.rs
@@ -1,4 +1,3 @@
 //! Types for EPP host responses
 
-pub mod info;
 pub mod update;
